Shuffle `format_exponential_notation()` methods
In `astropy.units` the `Base` unit formatter has three direct
subclasses: `CDS`, `Console` and `Generic`. Previously `Generic`
inherited its `format_exponential_notation()` method from `Base`, and
`CDS` and `Console` implemented their own, but their implementations
were very similar. With only very small changes it was possible to have
a common implementation for both `CDS` and `Console` in `Base`, meaning
`Generic` now has to have its own (very simple) implementation. This
does change how `Base.format_exponential_notation()` behaves, so it is
backwards-incompatible for any class that inherits directly from `Base`
without implementing its own `format_exponential_notation()` (or
implementing it using `super().format_exponential_notation()`), but
migration is very simple because any such class can replace
`Base.format_exponential_notation()` calls with calls of the built-in
`format()` function.
