+++ docs/changes/units/16676.api.rst
@@ -1,0 +1,16 @@
+The ``format_exponential_notation()`` method of the ``Base`` unit formatter has
+changed.
+Any unit formatters that inherit directly from ``Base`` but have not
+implemented their own ``format_exponential_notation()`` and wish to retain
+previous behavior should implement it as:
+
+.. code-block:: python
+
+    def format_exponential_notation(cls, val, format_spec):
+        return format(val, format_spec)
+
+Any formatters that inherit directly from ``Base`` and call
+``super().format_exponential_notation(val, format_spec)`` should instead call
+``format(val, format_spec)``
+The specific unit formatters in ``astropy.units`` and custom formatters that
+inherit from any of them are not affected.
